requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==0.29.37
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
editdistance==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpeg==1.4
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
funasr==0.8.8
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gradio==4.14.0
gradio_client==0.8.0
grpcio==1.60.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
loguru==0.7.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mecab-python3==1.0.8
modelscope==1.10.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.9.10
oss2==2.18.4
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.23.4
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pypinyin==0.50.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-wpe==0.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.35.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2+cu118
torch-complex==0.4.3
torchaudio==2.1.2
torchvision==0.16.2+cu118
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
umap==0.1.1
urllib3==2.1.0
uvicorn==0.25.0
websockets==11.0.3
Werkzeug==3.0.1
win32-setctime==1.1.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
